refactor: extract token encoder utils for zhipuai
