+++ libs/langchain-community/src/chat_models/zhipuai.ts
@@ -10,7 +10,8 @@ import {
 import { type ChatResult } from "@langchain/core/outputs";
 import { type CallbackManagerForLLMRun } from "@langchain/core/callbacks/manager";
 import { getEnvironmentVariable } from "@langchain/core/utils/env";
-import * as jwt from "jsonwebtoken";
+
+import { encodeApiKey } from "../utils/zhipuai.js"
 
 export type ZhipuMessageRole = "system" | "assistant" | "user";
 
@@ -190,7 +191,7 @@ export class ChatZhipuAI extends BaseChatModel implements ChatZhipuAIParams {
   constructor(fields: Partial<ChatZhipuAIParams> & BaseChatModelParams = {}) {
     super(fields);
 
-    this.zhipuAIApiKey = ChatZhipuAI.encodeApiKey(
+    this.zhipuAIApiKey = encodeApiKey(
       fields?.zhipuAIApiKey ?? getEnvironmentVariable("ZHIPUAI_API_KEY")
     );
     if (!this.zhipuAIApiKey) {
@@ -342,27 +343,6 @@ export class ChatZhipuAI extends BaseChatModel implements ChatZhipuAIParams {
         },
       },
     };
-  }
-
-  static encodeApiKey(apiKey: string | undefined) {
-    if (!apiKey) throw Error("Invalid api key");
-    const [key, secret] = apiKey.split(".");
-    const API_TOKEN_TTL_SECONDS = 3 * 60;
-    const now = new Date().valueOf();
-    const payload = {
-      api_key: key,
-      exp: now + API_TOKEN_TTL_SECONDS * 1000,
-      timestamp: now,
-    };
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const options: any = {
-      algorithm: "HS256",
-      header: {
-        alg: "HS256",
-        sign_type: "SIGN",
-      },
-    };
-    return jwt.sign(payload, secret, options);
   }
 
   /** @ignore */
